pkp/pkp-lib#3673 Fix unnecessary db lookup on all issues
This commit fixes a performance issue in which the cache was missed
for issue lookups. This happened because a PublishedArticle object
can be returned from PublishedArticleDAO without full submission
data. As a result, the IssueDAO::getById method was not receiving
an issue ID, and the cache was falling through each time to a series
of broader DB calls, culminating in a fetch of every single issue
in an attempt to identify it by a setting.

On large installations, this could create thousands of extra db
calls in a request.